perf(editor): defer browser PDF runtime
Split raster and PDF browser exporter entry points so editor startup and PNG export no longer load jsPDF. Make both human and Agent artifact paths import only the requested runtime while retaining the combined browser facade for compatibility.

Extend the production preview contract to reject an eagerly loaded browser-pdf chunk. Validated with workspace build, production smoke, BrowserAgentFileHost unit tests, and the formal Project/SVG/PNG/PDF browser export workflow. The gate plan selects full delivery for the shared exporter boundary; that remains required before merge.

Test-Impact: tests-updated
